+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 Newest updates are at the top of this file.
-## July 11 2018 - v4.0.8
+
+## July 22 2019 - v4.0.9
+* mqmetric - Support RESET QSTATS on z/OS queue manager
+* mqmetric - Add a Logger class to enable debug output
+* mqmetric - Improve some error reports
+
+## July 11 2019 - v4.0.8
 * Update for MQ 9.1.3 - No new API function introduced
 * mqmetric - Fix leak in subscriptions after rediscovery
 * mqmetric - Add USAGE as a queue label for selection by xmitq